requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiohttp-client-cache==0.6.1
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==21.4.0
Brotli==1.1.0
cattrs==23.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
HomeAssistant-API==4.2.1
idna==3.6
itsdangerous==2.1.2
multidict==6.0.5
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-forge==18.6.0
requests==2.31.0
requests-cache==0.9.8
simplejson==3.19.2
six==1.16.0
typing_extensions==4.9.0
url-normalize==1.4.3
urllib3==2.2.0
yarl==1.9.4
